+++ src/core/interfaces/album.interface.ts
