+++ src/core/interfaces/artist.interface.ts
