+++ src/core/interfaces/favorites.interface.ts
