+++ src/core/interfaces/track.interface.ts
