+++ src/core/interfaces/user.interface.ts
